modified:   .metadata/.log 	modified:   .metadata/.plugins/org.eclipse.e4.workbench/workbench.xmi 	new file:   biblioteca/mvc/controllers/PanelController.php 	new file:   biblioteca/mvc/controllers/TemaController.php 	modified:   biblioteca/mvc/views/libro/edit.php 	modified:   biblioteca/mvc/views/libro/show.php 	new file:   biblioteca/mvc/views/panel/panel.php 	modified:   biblioteca/templates/Base.php

+++ biblioteca/mvc/views/libro/edit.php
@@ -73,7 +73,9 @@
 			<a class="button" onclick="history.back()">Atrás</a>
 			<a class="button" href="/Libro/list">Lista de libros</a>
 			<a class="button" href="/Libro/show/<?=$libro->id?>">Detalles</a>
-			<a class="button-danger" href="/Libro/delete/<?=$libro->id?>">Borrado</a>
+			<?php if(!$libro->hasAny('Ejemplar')){ ?> 
+				<a class="button-danger" href="/Libro/delete/<?=$libro->id?>">Borrado</a>
+			<?php } ?>
 		</div>		
 	</form>
 </main>
